+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # 0.0.2.beta
 
+## Dialog Cancel button
+
+Dialogs have been updated to pass context to any children components. We have used this to switch the Form Cancel button to use the Dialog's cancel handler when the form is nested in a dialog. This overrides the default history.back method.
+
 ## Store, View and Route Utils
 
 We have standardised the utilities we provide to easily set up Flux based applications. This involved a few breaking changes:
@@ -110,6 +114,7 @@ A new style and colour scheme has been applied to the carbon components library.
 ## Misc
 
 * Ran ESLint task and fixed any errors.
+* Forms no longer needs a model name defined.
 * Updated Form Cancel Button to use History object.
 * Input components can now use custom classes.
 * Added Rainbow chart component.